requirements.txt
asyncio==3.4.3
GitPython==3.1.40
PyYAML==6.0.1
aiofiles==23.2.1
requests==2.31.0
psutil==5.9.6
redis==5.0.1
boto3==1.34.34
google-cloud-compute==1.15.0
azure-mgmt-compute==30.4.0
prometheus-client==0.19.0
dataclasses-json==0.6.3
